fix some issues + change directory name because _0base not valid name for this directory + BaseActivity and BaseViewModel are monsters. keep simple + create MessageManager.kt for show own message + in BaseViewModel.kt we don't need safe call
