--- app/src/main/java/com/vicky7230/tasker/ui/_0base/BaseActivity.kt
@@ -1,104 +1,0 @@
-package com.vicky7230.tasker.ui._0base
-
-import android.annotation.SuppressLint
-import android.app.Dialog
-import android.content.pm.PackageManager
-import android.os.Build
-import android.os.Bundle
-import android.widget.Toast
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.content.ContextCompat
-import androidx.lifecycle.Observer
-import com.google.android.material.snackbar.Snackbar
-import com.vicky7230.tasker.R
-import com.vicky7230.tasker.events.TokenExpireEvent
-import com.vicky7230.tasker.utils.CommonUtils
-import org.greenrobot.eventbus.EventBus
-import org.greenrobot.eventbus.Subscribe
-import org.greenrobot.eventbus.ThreadMode
-
-
-@SuppressLint("Registered")
-abstract class BaseActivity : AppCompatActivity() {
-
-    private var progressDialog: Dialog? = null
-    private lateinit var baseViewModel: BaseViewModel
-
-    abstract fun getViewModel(): BaseViewModel
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        baseViewModel = getViewModel()
-        baseViewModel.loggedOut.observe(this, Observer {
-            /*startActivity(LoginActivity.getStartIntentNewTask(this))
-            finish()*/
-        })
-    }
-
-    fun hasPermissions(permissions: Array<String>): Boolean {
-        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.M)
-            return true
-        return permissions.none {
-            ContextCompat.checkSelfPermission(this, it) != PackageManager.PERMISSION_GRANTED
-        }
-    }
-
-    fun requestPermissionsSafely(permissions: Array<String>, requestCode: Int) {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-            requestPermissions(permissions, requestCode)
-        }
-    }
-
-    private fun displayMessage(message: String) {
-        Snackbar.make(findViewById(android.R.id.content), message, Snackbar.LENGTH_SHORT)
-            .show()
-    }
-
-    private fun displayError(message: String) {
-        val snackBar =
-            Snackbar.make(findViewById(android.R.id.content), message, Snackbar.LENGTH_SHORT)
-        snackBar.view.setBackgroundResource(R.drawable.curved_bg_error)
-        snackBar.show()
-    }
-
-    fun showLoading() {
-        hideLoading()
-        progressDialog = CommonUtils.showLoadingDialog(this)
-    }
-
-    fun hideLoading() {
-        if (progressDialog != null) {
-            if (progressDialog?.isShowing == true)
-                progressDialog?.cancel()
-        }
-    }
-
-    fun showMessage(message: String?) {
-        if (message != null)
-            displayMessage(message)
-    }
-
-    fun showError(message: String?) {
-        if (message != null)
-            displayError(message)
-    }
-
-    fun showToast(message: String?) {
-        Toast.makeText(this, message, Toast.LENGTH_LONG).show()
-    }
-
-    override fun onStart() {
-        super.onStart()
-        EventBus.getDefault().register(this)
-    }
-
-    override fun onStop() {
-        EventBus.getDefault().unregister(this)
-        super.onStop()
-    }
-
-    @Subscribe(threadMode = ThreadMode.MAIN)
-    open fun onTokenExpire(event: TokenExpireEvent) {
-        baseViewModel.onTokenExpire()
-    }
-}
