+++ app/src/main/java/com/vicky7230/tasker/ui/_4home/HomeViewModel.kt
@@ -2,20 +2,13 @@ package com.vicky7230.tasker.ui._4home
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.viewModelScope
-import com.google.gson.JsonElement
-import com.google.gson.JsonObject
 import com.vicky7230.tasker.data.DataManager
-import com.vicky7230.tasker.data.db.entities.Task
 import com.vicky7230.tasker.data.db.entities.TaskList
 import com.vicky7230.tasker.data.db.joinReturnTypes.TaskAndTaskList
 import com.vicky7230.tasker.data.db.joinReturnTypes.TaskListAndCount
 import com.vicky7230.tasker.data.network.Resource
-import com.vicky7230.tasker.events.TokenExpireEvent
-import com.vicky7230.tasker.ui._0base.BaseViewModel
-import kotlinx.coroutines.flow.collect
+import com.vicky7230.tasker.ui.base.BaseViewModel
 import kotlinx.coroutines.launch
-import org.greenrobot.eventbus.EventBus
-import java.io.IOException
 import javax.inject.Inject
 
 class HomeViewModel @Inject constructor(
