+++ app/src/main/java/com/vicky7230/tasker/ui/_5newTask/NewTaskViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.viewModelScope
 import com.vicky7230.tasker.data.DataManager
 import com.vicky7230.tasker.data.db.entities.Task
 import com.vicky7230.tasker.data.db.entities.TaskList
-import com.vicky7230.tasker.ui._0base.BaseViewModel
+import com.vicky7230.tasker.ui.base.BaseViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
