+++ app/src/main/java/com/vicky7230/tasker/ui/_7finishedDeleted/FinishedDeletedTasksViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.viewModelScope
 import com.vicky7230.tasker.data.DataManager
 import com.vicky7230.tasker.data.db.joinReturnTypes.TaskAndTaskList
-import com.vicky7230.tasker.ui._0base.BaseViewModel
+import com.vicky7230.tasker.ui.base.BaseViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
