+++ app/src/main/java/com/vicky7230/tasker/ui/base/BaseActivity.kt
@@ -1,0 +1,16 @@
+package com.vicky7230.tasker.ui.base
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+
+abstract class BaseActivity : AppCompatActivity() {
+
+    private lateinit var baseViewModel: BaseViewModel
+
+    abstract fun getViewModel(): BaseViewModel
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        baseViewModel = getViewModel()
+    }
+}
