+++ app/src/main/java/com/vicky7230/tasker/ui/base/BaseViewModel.kt
@@ -1,0 +1,5 @@
+package com.vicky7230.tasker.ui.base
+
+import androidx.lifecycle.ViewModel
+
+open class BaseViewModel : ViewModel()
